move cloudnode to its own module (#339)
* move cloudnode to it's own module

* lint

* fix tests
