+++ fusillade/directory/__init__.py
@@ -5,7 +5,6 @@ This modules is used to simplify access to AWS Cloud Directory. For more informa
 https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/clouddirectory.html
 
 """
-import hashlib
 import json
 import logging
 import os
@@ -17,10 +16,13 @@ from dcplib.aws import clients as aws_clients
 from fusillade import Config
 from fusillade.config import proj_path
 from fusillade.directory.clouddirectory import CloudDirectory
-from fusillade.directory.structs import UpdateActions, obj_type_path, ValueTypes, ConsistencyLevel, UpdateObjectParams
-from fusillade.errors import FusilladeException, FusilladeHTTPException, FusilladeNotFoundException, \
-    AuthorizationException, FusilladeLimitException, FusilladeBadRequestException
+from fusillade.directory.cloudnode import CloudNode
+from fusillade.directory.structs import UpdateActions, ValueTypes, ConsistencyLevel, UpdateObjectParams
+from fusillade.directory.identifiers import obj_type_path, get_obj_type_path
+from fusillade.errors import FusilladeHTTPException, FusilladeNotFoundException, \
+    AuthorizationException, FusilladeLimitException
 from fusillade.policy.validator import verify_policy
+from fusillade.utils.json import get_json_file
 
 logger = logging.getLogger(__name__)
 
@@ -36,11 +38,6 @@ default_group_policy_path = os.path.join(proj_path, '..', 'policies', 'default_g
 default_admin_role_path = os.path.join(proj_path, '..', 'policies', 'default_admin_role.json')
 default_user_role_path = os.path.join(proj_path, '..', 'policies', 'default_user_role.json')
 default_role_path = os.path.join(proj_path, '..', 'policies', 'default_role.json')
-
-
-def get_json_file(file_name) -> Dict[str, Any]:
-    with open(file_name, 'r') as fp:
-        return json.load(fp)
 
 
 def get_published_schema_from_directory(dir_arn: str) -> str:
@@ -181,263 +178,6 @@ def clear_cd(directory: 'CloudDirectory',
         directory.delete_object(obj_ref, delete_children=True)
 
 
-class CloudNode:
-    """
-    Contains shared code across the different types of nodes stored in Fusillade CloudDirectory
-    """
-    _attributes = ["name"]  # the different attributes of a node stored
-    _facet = 'LeafFacet'
-    object_type = 'node'
-
-    def __init__(self,
-                 name: str = None,
-                 object_ref: str = None):
-        """
-
-        :param name:
-        :param object_ref:
-        """
-        self.cd: CloudDirectory = Config.get_directory()
-        if name and object_ref:
-            raise FusilladeException("object_reference XOR name")
-        if name:
-            self.from_name(name)
-        else:
-            self._name: str = None
-            self._path_name: str = None
-            self.object_ref: str = object_ref
-        self.attached_policies: Dict[str, str] = dict()
-
-    def from_name(self, name):
-        self._name: str = name
-        self._path_name: str = self.hash_name(name)
-        self.object_ref: str = self.cd.get_obj_type_path(self.object_type) + self._path_name
-
-    @staticmethod
-    def hash_name(name):
-        """Generate the cloud directory path name from the nodes name."""
-        return hashlib.sha1(bytes(name, "utf-8")).hexdigest()
-
-    def _get_link_name(self, parent_path: str, child_path: str):
-        return self.hash_name(parent_path + child_path)
-        # links names must be unique between two objects
-
-    def _get_links(self, node: Type['CloudNode'],
-                   filter_attribute_range: List[Dict[str, Any]],
-                   facet,
-                   next_token=None,
-                   per_page=None,
-                   paged=False,
-                   incoming=False):
-        """
-        Retrieves the links attached to this object from CloudDirectory and separates them into groups and roles
-        based on the link name
-        """
-        get_links = self.cd.list_incoming_typed_links if incoming else self.cd.list_outgoing_typed_links
-        object_selection = 'SourceObjectReference' if incoming else 'TargetObjectReference'
-        if paged:
-            result, next_token = get_links(self.object_ref, filter_attribute_range, facet,
-                                           next_token=next_token, paged=paged, per_page=per_page)
-            if result:
-                operations = [self.cd.batch_get_attributes(
-                    obj_ref[object_selection]['Selector'],
-                    node._facet,
-                    ['name'])
-                    for obj_ref in result]
-                result = []
-                for r in self.cd.batch_read(operations)['Responses']:
-                    if r.get('SuccessfulResponse'):
-                        result.append(
-                            r.get('SuccessfulResponse')['GetObjectAttributes']['Attributes'][0]['Value']['StringValue'])
-                    else:
-                        logger.error({"message": "Batch Request Failed", "response": r})  # log error request failed
-            return result, next_token
-        else:
-            return [
-                type_link[object_selection]['Selector']
-                for type_link in
-                get_links(self.object_ref, filter_attribute_range, facet)
-            ]
-
-    def _add_links_batch(self, links: List[Type['CloudNode']]):
-        """
-        Attaches links to this object in CloudDirectory.
-        """
-        if not links:
-            return []
-        batch_attach_object = self.cd.batch_attach_object
-        operations = []
-        for link in links:
-            parent_ref = link.object_ref
-            operations.append(
-                batch_attach_object(
-                    parent_ref,
-                    self.object_ref,
-                    self._get_link_name(parent_ref, self.object_ref)
-                )
-            )
-        return operations
-
-    def _add_typed_links_batch(self, links: List[Type['CloudNode']], link_type: str, attributes: Dict, incoming=False):
-        """
-        Attaches links to this object in CloudDirectory.
-
-        TODO modify this function to take in links: List[Type['CloudNode'] and remove the object_type parameter.
-        """
-        if not links:
-            return []
-        batch_attach_typed_link = self.cd.batch_attach_typed_link
-        operations = []
-        for link in links:
-            if incoming:
-                source, target = link.object_ref, self.object_ref
-            else:
-                source, target = self.object_ref, link.object_ref
-            operations.append(batch_attach_typed_link(source, target, link_type, attributes))
-        return operations
-
-    def _remove_links_batch(self, links: List[Type['CloudNode']], incoming=False):
-        """
-        Removes links from this object in CloudDirectory.
-        """
-        if not links:
-            return []
-        batch_detach_object = self.cd.batch_detach_object
-        operations = []
-        for link in links:
-            if incoming:
-                source, target = link.object_ref, self.object_ref
-            else:
-                source, target = self.object_ref, link.object_ref
-            operations.append(
-                batch_detach_object(
-                    target,
-                    self._get_link_name(target, source)
-                )
-            )
-        return operations
-
-    def _remove_typed_links_batch(self, links: List[Type['CloudNode']], link_type: str, attributes: Dict,
-                                  incoming=False):
-        """
-        Removes links from this object in CloudDirectory.
-        """
-        if not links:
-            return []
-        batch_detach_typed_link = self.cd.batch_detach_typed_link
-        make_typed_link_specifier = self.cd.make_typed_link_specifier
-        operations = []
-        for link in links:
-            if incoming:
-                source, target = link.object_ref, self.object_ref
-            else:
-                source, target = self.object_ref, link.object_ref
-            typed_link_specifier = make_typed_link_specifier(
-                source,
-                target,
-                link_type,
-                attributes
-            )
-            operations.append(batch_detach_typed_link(typed_link_specifier))
-        return operations
-
-    @property
-    def name(self):
-        if not self._name:
-            self._get_attributes(self._attributes)
-            self._path_name = self.hash_name(self._name)
-        return self._name
-
-    def _get_attributes(self, attributes: List[str]):
-        """
-        retrieve attributes for this from CloudDirectory and sets local private variables.
-        """
-        if not attributes:
-            attributes = self._attributes
-        try:
-            resp = self.cd.get_object_attributes(self.object_ref, self._facet, attributes)
-        except cd_client.exceptions.ResourceNotFoundException:
-            raise FusilladeNotFoundException(detail="Resource does not exist.")
-        for attr in resp['Attributes']:
-            self.__setattr__('_' + attr['Key']['Name'], attr['Value'].popitem()[1])
-
-    def get_attributes(self, attributes: List[str]) -> Dict[str, str]:
-        try:
-            resp = self.cd.get_object_attributes(self.object_ref, self._facet, attributes)
-        except cd_client.exceptions.ResourceNotFoundException:
-            raise FusilladeNotFoundException(detail="Resource does not exist.")
-        return dict([(attr['Key']['Name'], attr['Value'].popitem()[1]) for attr in resp['Attributes']])
-
-    def delete_node(self):
-        try:
-            self.cd.delete_object(self.object_ref)
-        except cd_client.exceptions.ResourceNotFoundException:
-            raise FusilladeNotFoundException(f"Failed to delete {self.name}. {self.object_type} does not exist.")
-
-    @classmethod
-    def list_all(cls, next_token: str = None, per_page: int = None):
-        cd = Config.get_directory()
-        resp, next_token = cd.list_object_children_paged(f'/{cls.object_type}/', next_token, per_page)
-        operations = [cd.batch_get_attributes(f'${obj_ref}', cls._facet, ['name'])
-                      for obj_ref in resp.values()]
-        results = []
-        for r in cd.batch_read(operations)['Responses']:
-            if r.get('SuccessfulResponse'):
-                results.append(
-                    r.get('SuccessfulResponse')['GetObjectAttributes']['Attributes'][0]['Value']['StringValue'])
-            else:
-                logger.error({"message": "Batch Request Failed", "response": r})  # log error request failed
-        return {f"{cls.object_type}s": results}, next_token
-
-    def get_info(self) -> Dict[str, Any]:
-        info = dict(**self.get_attributes(self._attributes))
-        info[f'{self.object_type}_id'] = info.pop('name')
-        return info
-
-    @classmethod
-    def _exists(cls, nodes: List[str]):
-        operations = []
-        directory = Config.get_directory()
-        try:
-            for node in nodes:
-                operations.append(directory.batch_get_object_info(cls(node).object_ref))
-            directory.batch_read(operations)
-        except cd_client.exceptions.ResourceNotFoundException:
-            raise FusilladeBadRequestException(f"One or more {cls.object_type} does not exist.")
-
-    @classmethod
-    def get_names(cls, obj_refs: List[str]) -> List[str]:
-        cd = Config.get_directory()
-        operations = [cd.batch_get_attributes(obj_ref, cls._facet, ['name']) for obj_ref in obj_refs]
-        results = []
-        for r in cd.batch_read(operations)['Responses']:
-            if r.get('SuccessfulResponse'):
-                results.append(
-                    r.get('SuccessfulResponse')['GetObjectAttributes']['Attributes'][0]['Value']['StringValue'])
-            else:
-                logger.error({"message": "Batch Request Failed", "response": r})  # log error request failed
-        return results
-
-    def list_owners(self, incoming=True):
-        get_links = self.cd.list_incoming_typed_links if incoming else self.cd.list_outgoing_typed_links
-        object_selection = 'SourceObjectReference' if incoming else 'TargetObjectReference'
-        _owners = [
-            type_link[object_selection]['Selector']
-            for type_link in
-            get_links(self.object_ref, filter_typed_link='ownership_link')
-        ]
-        owners = []
-        for owner in _owners:
-            node = CloudNode(object_ref=owner)
-            owners.append({
-                'type': [i for i in
-                         [p['Path'].split('/')[1]
-                          for p in node.cd.list_object_parent_paths(node.object_ref)] if i != 'role'][0],
-                'name': node.name
-            })
-        return owners
-
-
 class PolicyMixin:
     """Adds policy support to a cloudNode"""
     allowed_policy_types = ['IAMPolicy']
@@ -459,7 +199,7 @@ class PolicyMixin:
         verify_policy(statement, policy_type)
         object_attribute_list = self.cd.get_policy_attribute_list('IAMPolicy', statement, **kwargs)
         policy_link_name = self.get_policy_name(policy_type)
-        parent_path = self.cd.get_obj_type_path('policy')
+        parent_path = get_obj_type_path('policy')
         operations.append(
             {
                 'CreateObject': {
@@ -508,7 +248,7 @@ class PolicyMixin:
         """
         if policy_type in self.allowed_policy_types:  # check if this policy type is allowed
             if not self.attached_policies.get(policy_type):  # check if we already have the policy
-                policy_ref = self.cd.get_obj_type_path('policy') + self.get_policy_name(policy_type)
+                policy_ref = get_obj_type_path('policy') + self.get_policy_name(policy_type)
                 try:
                     resp = self.cd.get_object_attributes(
                         policy_ref,
@@ -554,7 +294,7 @@ class PolicyMixin:
         ]
         try:
             try:
-                self.cd.update_object_attribute(self.cd.get_obj_type_path('policy') + self.get_policy_name(policy_type),
+                self.cd.update_object_attribute(get_obj_type_path('policy') + self.get_policy_name(policy_type),
                                                 params,
                                                 self.cd.node_schema)
             except cd_client.exceptions.ResourceNotFoundException:
@@ -591,7 +331,7 @@ class CreateMixin(PolicyMixin):
         new_node = cls(name)
         _creator = creator if creator else "fusillade"
         ops.append(new_node.cd.batch_create_object(
-            new_node.cd.get_obj_type_path(cls.object_type),
+            get_obj_type_path(cls.object_type),
             new_node.hash_name(name),
             new_node._facet,
             new_node.cd.get_object_attribute_list(facet=new_node._facet, name=name, created_by=_creator, **kwargs)
