+++ tests/test_groups.py
@@ -6,8 +6,9 @@ pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noq
 sys.path.insert(0, pkg_root)  # noqa
 
 from fusillade.errors import FusilladeHTTPException
-from fusillade.directory import User, Group, cd_client, cleanup_directory, cleanup_schema, get_json_file, \
+from fusillade.directory import User, Group, cd_client, cleanup_directory, cleanup_schema, \
     default_group_policy_path, Role, clear_cd
+from fusillade.utils.json import get_json_file
 from tests.common import new_test_directory, create_test_IAMPolicy, normalize_json
 from tests.infra.testmode import standalone
 from tests.json_mixin import AssertJSONMixin
