+++ tests/test_users.py
@@ -8,6 +8,7 @@ sys.path.insert(0, pkg_root)  # noqa
 from fusillade.errors import FusilladeHTTPException
 from fusillade.directory import User, Group, Role, cd_client, cleanup_directory, cleanup_schema, \
     get_json_file, default_user_policy_path, default_user_role_path, default_group_policy_path, clear_cd
+from fusillade.utils.json import get_json_file
 from tests.common import new_test_directory, create_test_IAMPolicy, normalize_json
 from tests.infra.testmode import standalone
 from tests.json_mixin import AssertJSONMixin
